webapi/webapi-webrtc-w3c-tests/webrtc/blink/RTCPeerConnection-remoteDescription.html

<!DOCTYPE html>
<!--
Test converted from blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/mediastream/RTCPeerConnection-remoteDescription.html
-->

<meta charset=utf-8>
<title>WebRTC Test: RTCPeerConnection remoteDescription</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel=help href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#widl-RTCPeerConnection-setRemoteDescription-void-RTCSessionDescription-description-VoidFunction-successCallback-RTCPeerConnectionErrorCallback-failureCallback">
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<div id=log></div>
<script>

var pc = null;

var t = async_test(document.title);

function requestFailed2() {
  t.step(function () {
    assert_equals(pc.remoteDescription.type, "answer");
    assert_equals(pc.remoteDescription.sdp, "remote");

    pc.close();
    assert_equals(pc.remoteDescription.type, "answer");
    assert_equals(pc.remoteDescription.sdp, "remote");
  });
  t.done();
}

function requestSucceeded2() {
  t.step(function () {
    assert_unreached("requestSucceeded was called.");
  });
  t.done();
}

function requestFailed1() {
  t.step(function () {
    assert_unreached("requestFailed was called.");
  });
  t.done();
}

function requestSucceeded1() {
  sessionDescription = new RTCSessionDescription({type:"offer", sdp:"local"});
  t.step(function () {
    pc.setRemoteDescription(sessionDescription, requestSucceeded2, requestFailed2);
  });
}

pc = new webkitRTCPeerConnection(null, null);
var sessionDescription = new RTCSessionDescription({type:"answer", sdp:"remote"});

test(function() {
  pc.setRemoteDescription(sessionDescription, requestSucceeded1, requestFailed1);
}, "Test RTCPeerConnection.setRemoteDescription()");

</script>
